requirements.txt
wheel
streamlit
pytest
ipykernel
yfinance
pandas-datareader
numpy-financial
plotly
# pandas